requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
python-dotenv
volcengine
pywinauto
pillow
pytesseract
selenium>=4.15.0
